long trocado para int

+++ Projeto-1/projeto_1-zip/TAD/DNA.h
@@ -7,11 +7,11 @@ typedef struct {
 	FILE *genoma;
 
 	FILE *pos_genes;
-	long n_genes;
+	int n_genes;
 
 	FILE *fragmentos;
-	long **pos_fragmentos;
-	long n_fragmentos;
+	int **pos_fragmentos;
+	int n_fragmentos;
 
 }DNA;
 
@@ -19,19 +19,19 @@ DNA *criar_DNA(
     char *genoma, 
     char *pos_genes, 
     char *fragmentos, 
-    long n_genes, 
-    long n_fragmentos
+    int n_genes, 
+    int n_fragmentos
 );
 
 int destruir_DNA(DNA *dna);
 
-void OrdenaDigitos(long A[][2], long n, long posicao);
+void OrdenaDigitos(int A[][2], int n, int posicao);
 
-void OrdenaNumeros(long A[][2], long n);
+void OrdenaNumeros(int A[][2], int n);
 
-void ContagemIntersecoes(FILE *arquivo_A, long **B, long nA, long nB, FILE *arquivo_saida);
+void ContagemIntersecoes(FILE *arquivo_A, int **B, int nA, int nB, FILE *arquivo_saida);
 
-long **CtrlF(FILE *arquivo_texto, FILE *arquivo_trechos, long n_fragmentos);
+int **CtrlF(FILE *arquivo_texto, FILE *arquivo_trechos, int n_fragmentos);
 
 
 void ContagemLeituras(DNA *dna, FILE *saida);
